Removed ambiguities on SpinnerNumberModel value type
git-svn-id: https://svn.code.sf.net/p/sweethome3d/code/trunk/SweetHome3D@669 03e323e1-6c6c-46a5-9f71-c4ce5d8c6204

+++ src/com/eteks/sweethome3d/swing/HomeFurniturePanel.java
@@ -282,7 +282,7 @@ public class HomeFurniturePanel extends JPanel {
       if (angle == null) {
         this.angleSpinner.setValue(null);
       } else {
-        this.angleSpinner.setValue((Math.round(Math.toDegrees(angle)) + 360) % 360);
+        this.angleSpinner.setValue((int)(Math.round(Math.toDegrees(angle)) + 360) % 360);
       }
 
       Float width = firstPiece.getWidth();
